+++ src/main/java/us/stephenmitchell/recipe_rest/advice/StepNotFoundAdvice.java
@@ -1,18 +1,19 @@
-package us.stephenmitchell.recipe_rest.controller;
+package us.stephenmitchell.recipe_rest.advice;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
+import us.stephenmitchell.recipe_rest.exception.StepNotFoundException;
 
 @ControllerAdvice
-public class RecipeNotFoundAdvice {
+public class StepNotFoundAdvice {
 
     @ResponseBody
-    @ExceptionHandler(RecipeNotFoundException.class)
+    @ExceptionHandler(StepNotFoundException.class)
     @ResponseStatus(HttpStatus.NOT_FOUND)
-    String recipeNotFoundHandler(RecipeNotFoundException ex) {
+    String stepNotFoundHandler(StepNotFoundException ex) {
         return ex.getMessage();
     }
 }
