+++ src/main/java/us/stephenmitchell/recipe_rest/assembler/RecipeAssembler.java
@@ -1,6 +1,8 @@
-package us.stephenmitchell.recipe_rest.controller;
+package us.stephenmitchell.recipe_rest.assembler;
 
+import org.springframework.hateoas.server.mvc.WebMvcLinkBuilder;
 import org.springframework.stereotype.Component;
+import us.stephenmitchell.recipe_rest.controller.RecipeController;
 import us.stephenmitchell.recipe_rest.model.Recipe;
 
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.*;
@@ -13,7 +15,7 @@ public class RecipeAssembler implements RepresentationModelAssembler<Recipe,Enti
     @Override
     public EntityModel<Recipe> toModel(Recipe recipe) {
         return EntityModel.of(recipe, //
-                linkTo(methodOn(RecipeController.class).one(recipe.getId())).withSelfRel(),
+                WebMvcLinkBuilder.linkTo(methodOn(RecipeController.class).one(recipe.getId())).withSelfRel(),
                 linkTo(methodOn(RecipeController.class).all()).withRel("recipe_list"));// )
     }
 }
