+++ src/main/java/us/stephenmitchell/recipe_rest/exception/RecipeNotFoundException.java
@@ -1,4 +1,4 @@
-package us.stephenmitchell.recipe_rest.controller;
+package us.stephenmitchell.recipe_rest.exception;
 
 public class RecipeNotFoundException extends RuntimeException {
     public RecipeNotFoundException(Long id) {
