+++ src/main/java/us/stephenmitchell/recipe_rest/exception/StepNotFoundException.java
@@ -1,4 +1,4 @@
-package us.stephenmitchell.recipe_rest.controller;
+package us.stephenmitchell.recipe_rest.exception;
 
 public class StepNotFoundException extends RuntimeException {
     public StepNotFoundException(Long id) {
